+++ infra/provider/CHANGELOG.md
@@ -1,5 +1,17 @@
 # provider-azure
 
+## 0.0.5
+
+### Patch Changes
+
+- 222b6a2: Improve resource_name function with italynorth/westeurope values support for location parameter
+
+## 0.0.4
+
+### Patch Changes
+
+- 03edd1a: Add Tests for Provider and Function, fix a typo and add new resources for postgresql and function app
+
 ## 0.0.3
 
 ### Patch Changes